refactored tests to existing classes
